// src/enums/PokerHands.ts
export enum PokerHands {
  HighCard = "High Card",
  Pair = "Pair",
  TwoPair = "Two Pair",
  ThreeOfKind = "Three of a Kind",
  Straight = "Straight",
  Flush = "Flush",
  FullHouse = "Full House",
  FourOfKind = "Four of a Kind",
  StraightFlush = "Straight Flush",
  RoyalFlush = "Royal Flush",
}